Refactor: Move command aliases to config.js
I moved the `commandAliases` map from `core/commandRegistry.js` to `config.js` to centralize configuration.
